Revert "misc: always assign miscdevice to file->private_data in open()"
This reverts commit 32eca22180804f71b06b63fd29b72f58be8b3c47.

Changing core kernel code to operate in a different manner, without a
build-time breakage is tough to do and ensure you got it right.  There
are lots of problems popping up due to this change, so let's revert it
for now as it is not safe to merge to the tree at this point in time.

Cc: Martin Kepplinger <martink@posteo.de>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>